Added basic encryption methods, but no proper interface
Note - I spent a while trying to package the encryption bit as a namespace package but it proved difficult with the current setup. This method effectively accomplishes the same task, but avoids the problem of namespace packages altogether. Downloading a few extra bytes of code isn't a big deal here, but it may be one should the module continue to gain additional optional features. The rational behind splitting the encryption into an optional feature set lies in the fact that it requires non-default packages.